fixes #29 - change all sent/recv counters from int32 to int64 to prevent wraparound

+++ src/libnethogs.h
@@ -24,8 +24,8 @@ typedef struct NethogsMonitorRecord {
   int pid;
   uint32_t uid;
   const char *device_name;
-  uint32_t sent_bytes;
-  uint32_t recv_bytes;
+  uint64_t sent_bytes;
+  uint64_t recv_bytes;
   float sent_kbs;
   float recv_kbs;
 } NethogsMonitorRecord;
